requirements.txt
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
dweepy==0.3.0
idna==3.4
paho-mqtt==1.6.1
psycopg2==2.9.4
pycodestyle==2.9.1
pyserial==3.5
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
toml==0.10.2
urllib3==1.26.12
